- (dtucker) [progressmeter.c scp.c sftp-server.c] Use correct casts for    snprintf formats, fixes warnings on some 64 bit platforms.  Patch from    shaw at vranix.com, ok djm@

+++ scp.c
@@ -563,7 +563,7 @@ syserr:			run_err("%s: %s", name, strerror(errno));
 #define	FILEMODEMASK	(S_ISUID|S_ISGID|S_IRWXU|S_IRWXG|S_IRWXO)
 		snprintf(buf, sizeof buf, "C%04o %lld %s\n",
 		    (u_int) (stb.st_mode & FILEMODEMASK),
-		    (int64_t)stb.st_size, last);
+		    (long long)stb.st_size, last);
 		if (verbose_mode) {
 			fprintf(stderr, "Sending file modes: %s", buf);
 		}
